+++ src/Components/Widget/NewInputDisplay.tsx
@@ -23,7 +23,7 @@ export const NewInputDisplay = memo((props: {
 
   return (
     <View>
-      <Text h2
+      <Text h3
         style={{
           color: props.theme.font,
           paddingHorizontal: 10,